Move TimeSignal stuff to allegedb, normalize the API
